requirements.txt
pytube
pydub
openai-whisper
azure-search-documents
python-dotenv
openai
langchain
rich
pyautogen